Added Some Window Options
Added two 'window' options: 'use_dummy' and 'minimize'window'.

If 'use_dummy' is true, a dummy window is created off screen. This prevents windows from throttling the system when "out of focus" since the dummy window is set to be always on top.

If 'minimize_window' is true, the GUI minimizes once the system is running.